+++ app/register/email/_components/RegisterName.tsx
@@ -1,6 +1,5 @@
 import style from '../styles/registerName.module.scss'
-import BaseAvatar from '/public/assets/baseAvatar.svg'
-import NameForm from './NameForm'
+import NameForm from './NameRegister'
 import { Dispatch, SetStateAction } from 'react'
 import { IInputValues } from '@/app/_lib/validate'
 
@@ -17,20 +16,11 @@ export default function RegisterName({
 }: Props) {
   return (
     <>
-      <h2 className={style.title}>
-        함께할 이름을 <br />
-        적어주세요
-      </h2>
-      <div className={style.avatarDiv}>
-        <BaseAvatar />
-      </div>
-      <div className={style.formDiv}>
-        <NameForm
-          nextStep={nextStep}
-          userTotalDatas={userTotalDatas}
-          setUserTotalDatas={setUserTotalDatas}
-        />
-      </div>
+      <NameForm
+        nextStep={nextStep}
+        userTotalDatas={userTotalDatas}
+        setUserTotalDatas={setUserTotalDatas}
+      />
     </>
   )
 }
