+++ app/register/email/styles/birthRegister.module.scss
