--- app/register/email/styles/birthSelect.module.scss
@@ -1,16 +1,0 @@
-.select {
-  &::-ms-expand {
-    display: none;
-  }
-  -o-appearance: none;
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  appearance: none;
-  width: 60%;
-  margin-inline: auto;
-  outline: none;
-  border-radius: 12px;
-  border: 1px solid #d9d9d9;
-  padding-block: 15px;
-  text-align: center;
-}
