+++ app/register/email/styles/nameRegister.module.scss
@@ -23,3 +23,20 @@
   font-weight: var(--fw-medium);
   line-height: 15px;
 }
+
+.title {
+  margin-top: 49px;
+  line-height: 1.5;
+  color: #000;
+  font-size: 24px;
+  font-weight: var(--fw-bold);
+}
+
+.avatarDiv {
+  margin-block: 42px 24px;
+  text-align: center;
+}
+
+.formDiv {
+  margin-bottom: 117px;
+}
