--- app/register/email/styles/registerEmail.module.scss
@@ -1,25 +1,0 @@
-.title {
-  margin-top: 32px;
-  color: #222;
-  font-size: 20px;
-  font-weight: var(--fw-semiBold);
-}
-
-.formDiv {
-  margin-block: 32px 175px;
-}
-
-.description {
-  word-break: keep-all;
-  width: 65%;
-  margin-top: 12px;
-  margin-inline: auto;
-  color: #d9d9d9;
-  text-align: center;
-  font-size: 12px;
-  font-weight: var(--fw-medium);
-  span {
-    text-decoration-line: underline;
-    color: #b3b3b3;
-  }
-}
