--- app/register/email/styles/registerName.module.scss
@@ -1,16 +1,0 @@
-.title {
-  margin-top: 49px;
-  line-height: 1.5;
-  color: #000;
-  font-size: 24px;
-  font-weight: var(--fw-bold);
-}
-
-.avatarDiv {
-  margin-block: 42px 24px;
-  text-align: center;
-}
-
-.formDiv {
-  margin-bottom: 117px;
-}
